pyproject.toml
[build-system]
requires = [
  "pdm-backend"
]
build-backend = "pdm.backend"

[dependency-groups]
dev = [
  "pdm>=2.22.3",
  "pdm-bump>=0.9.10",
  "pre-commit>=4.1.0",
  "pytest>=8.3.4"
]

[project]
name = "pypi-docs-url"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
  {name = "Louis Maddox", email = "louismmx@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
  "click>=8.1.8",
  "pyyaml>=6.0.2",
  "requests>=2.32.3"
]

[project.scripts]
pypi-docs-url = "pypi_docs_url.cli:main"
